refactor getAllAnswers
