+++ eq-author/src/App/QuestionnaireDesignPage/MainNavigation/index.js
@@ -1,11 +1,12 @@
 import React, { useState } from "react";
 import styled from "styled-components";
 import PropTypes from "prop-types";
-import { flowRight, find } from "lodash";
+import { flowRight, find, get } from "lodash";
 
 import { withRouter } from "react-router-dom";
 import gql from "graphql-tag";
 import { useSubscription } from "react-apollo";
+import { useQuery } from "@apollo/react-hooks";
 
 import config from "config";
 import CustomPropTypes from "custom-prop-types";
@@ -41,10 +42,11 @@ import {
   buildSettingsPath,
 } from "utils/UrlUtils";
 
-import GET_ALL_ANSWERS from "../../qcodes/QCodesTable/graphql/getAllAnswers.graphql";
-import { useQuery } from "@apollo/react-hooks";
+import GET_ALL_ANSWERS from "../../../graphql/getAllAnswers.graphql";
 import Error from "components/Error";
 import Loading from "components/Loading";
+import { organiseAnswers, flattenAnswers } from "../../../utils/getAllAnswersFlatMap"
+
 
 const StyledMainNavigation = styled.div`
   color: ${colors.grey};
@@ -110,135 +112,10 @@ export const UnwrappedMainNavigation = props => {
     console.log('error', error);
     return <Error>Oops! Something went wrong</Error>;
   }
-  const removeHtml = html => html && html.replace(/(<([^>]+)>)/gi, "");
 
-  const organiseAnswers = sections => {
-    const questions = sections.reduce(
-      (acc, section) => [...acc, ...section.pages],
-      []
-    );
+  const { sections } = get(data, "questionnaire", []);
+  // const { sections } = data.questionnaire;
 
-    let answerRows = [];
-
-    for (const item of questions) {
-      const {
-        title,
-        alias,
-        answers,
-        confirmation,
-        summaryAnswers: calculatedSummary,
-      } = item;
-
-      if (answers) {
-        const extraCheck = answers.reduce((acc, item) => {
-          if (
-            item.hasOwnProperty("options") &&
-            item.options &&
-            item.type !== "radio"
-          ) {
-            const optionLabel = item.options.map(option => ({
-              ...option,
-              type: "CheckboxOption",
-              option: true,
-            }));
-
-            acc.push(...optionLabel);
-          }
-
-          if (
-            item.hasOwnProperty("mutuallyExclusiveOption") &&
-            item.mutuallyExclusiveOption
-          ) {
-            acc.push({
-              ...item.mutuallyExclusiveOption,
-              type: "MutuallyExclusiveOption",
-              option: true,
-            });
-          }
-          if (
-            item.hasOwnProperty("secondaryLabel") &&
-            item.hasOwnProperty("secondaryQCode") &&
-            item.secondaryLabel
-          ) {
-            acc.push({
-              id: item.id,
-              label: item.secondaryLabel,
-              qCode: item.secondaryQCode,
-              type: item.type,
-              secondary: true,
-            });
-          }
-          return acc;
-        }, []);
-
-        const answersAndOptions = [...answers, ...extraCheck];
-
-        answerRows.push({
-          title,
-          alias,
-          answers: answersAndOptions,
-        });
-      }
-
-      if (confirmation) {
-        const { id, title, alias, qCode, __typename: type } = confirmation;
-
-        answerRows.push({
-          title: title,
-          alias,
-          answers: [{ id, qCode, type }],
-        });
-      }
-
-      if (calculatedSummary && calculatedSummary.length) {
-        const {
-          id,
-          pageType: type,
-          alias,
-          title,
-          qCode,
-          totalTitle,
-          summaryAnswers,
-        } = item;
-
-        const label = removeHtml(totalTitle);
-
-        answerRows.push({
-          title,
-          alias,
-          answers: [{ id, type, qCode, label, summaryAnswers }],
-        });
-      }
-    }
-
-    return { answers: answerRows };
-  };
-
-  const flattenAnswers = data => {
-    const answers = data.reduce((acc, item) => {
-      const answer = item.answers.map((ans, index) => {
-        if (index > 0) {
-          return {
-            title: item.title,
-            alias: item.alias,
-            nested: true,
-            ...ans,
-          };
-        } else {
-          return {
-            title: item.title,
-            alias: item.alias,
-            ...ans,
-          };
-        }
-      });
-      acc.push(...answer);
-      return acc;
-    }, []);
-    return answers;
-  };
-
-  const { sections } = data.questionnaire;
   const { answers } = organiseAnswers(sections);
   const flatten = flattenAnswers(answers);
