+++ eq-author/src/graphql/getAllAnswers.graphql
